requirements.txt
streamlit
python-dotenv
streamlit-oauth
requests
openai
bcrypt
PyPDF2